Rework how we set the default logging-config.
If we want to respect the environment variable before the global
default, we can't use the defaultConfigValues because they get set
before we have a chance to evaluate if the config has been set to use
the environment variable.
So most changes are back to what they were before the patch, just the
one line that use loggo's default config now uses our own default
config.