Include per-endpoint expose settings in exported bundles
If no per-endpoint expose settings have been specified, or a single
expose entry for the wildcard ("") endpoint exists that only contains a
0.0.0.0/0 CIDR (i.e. no expose-to-space-ids values), then the output
will match what a pre 2.9 controller would emit: the "exposed" flag will
be set to true.

On the other hand, if we do have expose settings present, the exposed
flag will be suppressed and the expose settings will be included as part
of the overlay part of the exported bundle.

This is an intentional move to prevent operators from accidentally
exposing their entire application to 0.0.0.0/0 if they export a bundle
from a 2.9 controller with expose settings and attempt to deploy it to a
2.8 (or earlier) controller that does not recognize the expose settings
section. By omitting the expose flag, we prevent the application from
being exposed at all and provide the operator with an opportunity to
audit the expose change and apply it manually if required.