Add firewaller API to fetch an optionally filtered list of space infos
In a similar fashion to the From/ToNetworkPortRange bits in the params
package, the conversion logic between a network.SpaceInfos and its
params equivalent is encapsulated in the params package. This makes it
quite easy to implement the same method in different facades if needed
without having to use networkingcommon.