Fix wordpress bundle
In order to use the wordpress bundle for manual tests, we need to fix
the unit tests to correctly handle the mysql error.

The underlying problem here is that we use real applications in bundles
and expect them to work for unit tests, integration tests and manual
tests. Instead we should probably have bundle yamls with unique software
names, so that they never collide with existing software in the wild.