Fix bug in handling of ingress rules in dummy provider's inst. firewall
The ingress diff logic in the firewall worker emits two sets of rules:
the open set (port + CIDRs not present in environ state) and the close
set (port + CIDRs present in environ state but not the firewaller's
state).

The implementation of the instance firewall Open/ClosePorts methods by
the dummy provider did not correcly handle processing of ingress rules
when the CIDR lists contained a subset of the CIDRs tracked by the dummy
provider.

To this end, the provider's logic has been updated to be aligned to how
the other providers operate. More specifically, when opening ports, the
provider will now match incoming rules to existing rules via a port
range check and then update the CIDR list of the existing rule to the
union of the existing and incoming CIDRs.

When closing ports, the provider will once again match the rules to be
removed to existing rules via a port range and then remove the CIDRs
from the rule to be removed from the matched rule's CIDR list. If the
existing rule's CIDR list is empty, the entire rule is discarded.