Removed deprecated firewaller facade API methods
As the firewall worker runs on the controller, it will switch to the new
API methods once the controller gets upgraded. As a result, we can
safely remove the following, now unused, methods without having to worry
about backwards compatibility issues: GetExposed,
GetMachineActiveSubnets, GetMachinePorts.

The removals and the updated facade schema are contained within this
commit and they can easily be reverted in the future if the need ever
arises.