Skip "live" tests for joyent provider as it is about to be removed
The firewall implementation for the joyent provider does not support
ingress rules that contain multiple source CIDRs and cause one of the
tests introduced by the previous commits into the LiveTest suite to
fail. As the provider will most probably be removed for 2.9, we can
safely skip the live tests for it.