Handle errors in charmhub client
This allows the handling of charmhub errors correctly in the charmhub
client. It allows the user of the result to combine the errors into one,
so we can pass it on to the user.

If there is no error in the api error list, we return nil and handle
that correctly.