Remove joyent from index.sjson used by the tests
The original index.sjson was signed by "Ubuntu Cloud Image Builder"
which means we cannot regenerate its signature after removing the joyent
entries given that we don't have access to these keys.

To this end, this commit removes the joyent bits from index.sjson and
regenerates its signature using the SignedMetadataPrivateKey contents
from environs/simplestreams/testing/testing.go. Unfortunately, this
change breaks the signature validation as the signer is not trusted. To
work around this, the commit extracts the signature check function into
a variable that we can patch as part of the tests which still runs the
original openpgp signature checker but suppresses ErrUnknownIssuer
errors.