Application status shouldn't use model cache
This fix is very subtle, because it highlights a problem with how the
all watcher is consumed by different clients (model cache, pylibjuju,
jslibjuju).

When you connect to the allwatcher you're given a series of deltas and
is expected that the client consumes that as verbatim. There is a
problem though, with the insert of the previous derviation status it
causes problems where it's impossible to know what the status of an
application is right now.

At the application delta is dispatched once at the start of signing on
to the allwatcher, you get what it was at the time of sign on, which is
inconsisent.

With the changes to getting the application status from the model cache
then depends if the units have primed the model cache in time to get
that information.

If we assume the diagram below we can see that depending on when you
join the allwatcher defines what application status.

If you're AllWatcher (A, B, C) you'll never see that the unit status is
blocked.

In order for pylibjuju the *only* way to get the right information is to
do it in pylibjuju.

```
 ALLWATCHER  ALLWATCHER  ALLWATCHER
     (A)         (B)         (C)
      +           +           +
      |           |           |
      |           |           |
      |           |           |
      |           |           |
      |           |           |
---+--+-----------+-+---------+------+-----------> TIME
   ^                ^                ^
   |                |                |
   |                |                |
   |                |                |
   +                +                +

APP STATUS      UNIT STATUS      UNIT STATUS
 (UNSET)         (WAITING)        (BLOCKED)
```