Ensure that we resolve a charm URL in a bundle
The following ensure that we correctly resolve the charm URL in a
bundle, if the bundle contains a non-canonical URL.

For example the following is a bad bundle:

```yaml
applications:
  wordpress:
      charm: wordpress
```

Instead it should be:

```yaml
applications:
  wordpress:
      charm: cs:wordpress
```

That way we always know that the wordpress correctly comes from the
right store. The code is simple and works with the charm hub
intergration.