Also include IPV6 wildcard CIDR when exposing applications to the world
This change allows us to be more IPV6-friendly by including the IPV6
wildcard CIDR (::/0) in addition to its IPV4 counterpart in the places
where the latter is added implicitly for the purpose of ingress rule
generation.

More specifically, this occurs:
- As part of the upgrade step for adding exposed endpoint entries to
applications exposed in older controllers.
- When the caller of the expose API omits endpoint specific entries
(e.g. juju expose X) OR requests for individual endpoints to be exposed
without providing any CIDRs or spaces (e.g. juju expose X --endpoints
Y).

Note that this expansion will not occur automatically when the operator
specifies a CIDR (even if it is the 0.0.0.0/0 one) when exposing an
application (or any of its endpoints). In this case we assume the
operator to know what they are doing and expect them to specify both
IPV4 and IPV6 CIDRs if that's required by their particular deployment
scenario.