Handle no application info for workload status
Although theoretically impossible in a normal workload setup, it has
shown up in tests. The idea that an application wouldn't exist, but have
a unit seems very implausible.
How ever implausible it's shown up in testing and if it can happen in
testing, it can show up in the wild.

The fix is really simple, if no application exists, we can safely ignore
the workload status and move on.