Ensure we use a ambigous CharmID
The CharmID that we currently have is banded about everywhere, which is
great, except that it comes from a third party package and it's tied to
the charmstore.

This move attempts to abolish it from the deploy stage, it still lingers
around in the resource deployments, but it's not far off from being
killed.