Explain why an application name is not valid
The following changes are rather trivial, but the idea is to explain to
the user why an application name isn't valid. It attempts to do this in
plain english where possible.

Rather than falling back to just telling the user about an unexpected
character, it first explains that uppercase characters aren't valid,
then numbers aren't allowed after hypens. Then if there is something
else wrong, explain which character is at fault.

Hopefully this will explain the real issue to the user when attempting
to interacting with juju.