Remove charmURL pointer map reference lookups
In the bundle handler code, there are a series of look up maps that
cache certain properties for a given bundle. Either that can be a charm
origin or a macaroon. Using a pointer for a lookup is wrong, because it
doesn't give you the expected result set if you're doing lookups.

```
a := charm.MustParseURL("wordpress")
b := charm.MustParseURL("wordpress")
a.Equals(b) // true
```

Variables a and b will probably not be the same underlying pointer, even
though they have the same representation.

This PR fixes that by ensure that we use the de-refrenced charm URL, so
that comparing charm url as values will get what you want. In fact if
you're using pointers in a map, you're probably doing it wrong and we
should probably have a static analysis tool for that.