Ensure we handle reuse of charms
Upgrading a bundle allows for us to reuse existing charms from previous
deployments and so means that we need to resolve the charm origin at the
correct time.

An example of the reuse can be:

```
$ juju deploy bundle.yaml

applications:
    django:
        charm: cs:bionic/django-42
        num_units: 1
```

The redeployment of the bundle with the node application pointing to the
django charm (troll 100%) should work.

```
$ juju deploy bundle.yaml

applications:
    django:
        charm: cs:bionic/django-42
        num_units: 1
    node:
        charm: cs:bionic/django-42
        num_units: 1
```

The code itself just stores all the charm origins when we resolve the
charms from the spec. This wasn't so obvious when hunting this down, but
in the end the fix is rather simple.