Query LXD network API to detect if space/subnet discovery should be supported
Some arcane lxd versions that ship for example with xenial/ppc64 do not
even support the "network" API extension. As a result, subnet discovery
cannot be enabled for those deployments.

This commit updates the SupportsSpaces and Subnets implementations of
the lxd networking interface to first check whether that particular
extension is supported by the LXD server and bail out if that's not the
case.