Ensure we use a valid channel for charm origin
Using the charm origin from the original command to help get the correct
charm. Even though the channel isn't 100% required as I don't believe
it's used further in the code, it doesn't help to be right in terms of
the channel.

I've also added tests for getting the core channel from a common api
charm.