Download charm from charm hub
The following implements a command for downloading a charm from a
charmhub store. The command follows simple unix terminology, by allowing
you to stream/pipe the download to a file. If you want the CLI to manage
the file handling, then supplying a --path will also do that for you.

The command allows you to talk to an existing bootstrapped model and get
that model config, with the charmhub url. If you want to specify your
own without a bootstrapped juju, then that is also possible with the
charm-hub-url path.

This command should work with both air gapped and non-air gapped
solutions.

The code itself required some changes, as we don't actually want to read
the charm archive once it's downloaded. Changing that so Download
becomes DownloadAndRead, which uses Download directly makes this more
usable.