Ensure we don't need a bootstrapped controller
The following code is a bit annoying. Essentially we want a model
command when we know we want to talk to a controller, but we also don't
want a model command when we don't need a controller.
The problem is that we don't know if we need a controller until we parse
the arguments. That is way too late and so we need to initialise a
model in a lazy fashion.

I'm not exactly happy with this work around, but without changing the
whole command type structure we are where we are. Interestingling we
could make this simpler if bootstrap, destroy, show and download didn't
inherit from the model command, but some other type that gave you
everything you needed, but not the model. I think the abstraction is at
the wrong level and an overhal of the juju command type is well overdue.