Pin the transent dependency (2.8.4)
The following dependency google_auth changes what it installs
depending on the version of python you have. I'm unsure of the real fix
here, as multidict doesn't seem to compile on anything by amd64. So
pinning it to use an older version seems a bit bruteforce'ish and we're
not getting the latest spec/security changes from that lib.

I think someone needs to come through and audit what is being installed
here and correctly pin them so we don't get blocked on a release.