Tighten up PR template a bit
Mainly moving the "description of change" section to the top, as (IMO) that's bar far the most important section. It will also improve the PR snippet/widget when you paste a PR link into GitHub or Trello, rather than having the checklist text in the way. I've also made some minor tweaks to the text/punctuation for consistency.