Add integer comparison
Before I forget how to do this and so have to do it at a later data,
I'll add integer comparison to make it easier in the long run.

Because of this change we have to do some things rather different. The
idea is that we force the GetIdentValue to return an Ord type and from
that we can do comparsion checks based on a type. We can then drop a lot
of the reflect value types. This makes the code much simpler in the long
run.