Allow unit queries
This allows for unit queries, but whilst testing unit querying I found
out that we drop the error if you misstype a identifier in a query and
this is no way to know why. The following code fixes that and as a bonus
we can check to see what you wanted as well.

The code is rather simple and quite a bit of copy and paste from a model
query. The difference being that the scope and the text for the command.