Pipe to stdout
Allow a charm to be piped to stdout if an addition hyphen argument is
passed. This follows the same pattern as config and other unix type
commands.

The code is really simple as we can use the FileSystem interface to
inject a new *os.File that points to the stdout pointer. Using that then
allows us to use the same machinary for downloading a charm that we've
always had.