Download a charm by either the channel or series
The following allows you to download a charm via a channel or series.
It's rather complicated to work out the right series by default and how
to correctly order them. Essentially we needed to order the controller
series (which essentially is all ubuntu releases) by the order of the
release (focal, bionic, xenial, etc). Then we order the resulting
channel map so that it aligns with that ordering, so that if no series
is defined, then you get an expected outcome.

There is currently no tests for that, those come later.