Map queries
This is the initial idea for adding map queries. The code introduces a
new set of tokens (left and right bracket) and we parse the index ast as
a infix operator. From there the query evaluator can just loop through
until we get to the bottom of the ast and collapse up.

The following are pefectly plausable:

    config["value"] == 1
    config["values"]["nested"] == 1

Slices are not currently handled, that will come later and also checking
the length of something is not correctly handled.