Rework run function for bash tests
There is a tension between running tests and running linters. The fix
previously worked great for linters, but failed hard for tests. This
fixes both paths.

We now have a run_linter command that turns out the ability to not have
to fail hard when you encounter an error, but run will continue to do
so.