Handle pipefail correctly
Allow pipefail to run for linters, but not for other tests. We need to
correctly handle pipefail to ensure we have the correct modes for the
tests.

This happens because we have a tension between the expected output from
the different running items.

The real fix is to use the fact that the subshell can fail, but we don't
want the subshell to bring down the parent shell. Correctly doing this
is kinda tricky, but I believe this is working as expected now.