Python3 changes in the acceptance tests
The following changes are required to get python3.6 or greater to work
with assess_upgrade.py.

This is a bit clumsy as I don't really know why we're getting a array of
bytes and not a string and probably should be ensuring that we
encode/decode into utf-8 in the correct places.