Define a run linter
Turns out the tension between run and run_linter was too great. This
instead makes it more evident that the two should never be merged
together as they require different things from the set modes.

Some time should be used to potentially merge the run functions together
with maybe a MODE envar, but for now we'll have duplicate functions
so we can make any tweaks without breaking the abstractions.