Allow function calls
This now allows function calls to happen in query. We currently only
support the len() built in function atm.

There is quite a few jump throughs to get len to work and jump down into
reflect, which isn't ideal, but it is what it is. You could probably use
a JIT here to speed up the caching of a function call, but if you're
going that far, then wow we got some time on our hands!