Ensure that logical AND and OR do it lazily
So when using AND and OR they should do things lazily and the right hand
side of the AND/OR should not be invoked unless the left hand side
passes. That way you can do things like:

    len(config) > 0 && config["name"] == "fred"

If you don't have lazy operators then it would execute the right hand
side and fail because name might not be there in the config yet.