Start to implement bootstrap simplestream test
The following starts to add a very simple stream server. It just
works in a very simple way by using the jujud available to it. Using
different tools to get this working easily (server, templating) makes
this much more efficent.

It tests that you can bootstrap with a simplestream server and then it
deploys onto that server.

Added a README file to help explain how the simplestream server works.
It is expected that the simplestream server is only used for a limited
set of tests and not a fully operational simplestream server to use in
any production scenarios.

---

Again the test is simple, put up a server, ensure that the tools are
there and bootstrap to that tool against the local simplestream server.