Start to implement upgrade bash testing
The following starts to add a very canned simple stream server. It just
works in a very simple way by using the jujud available to it. Using
different tools to get this working easily (server, templating) makes
this much more efficent.

Anyway we can bootstrap a controller using the supplied jujud, we now
have to work out to upgrade the same controller using the streams.