Make has containers allowed for advanceLifecycle to dying.
For cases like destroy-model and destroy-controller, allow for machines
to be moved to dying even if they have containers.  Containers must be
removed for a machine to be marked dead.  Use new DestroyWithContainers.

cleanupMachinesForDyingModel uses DestroyWithContainers, and not skip
containers, they need to be removed as well.

Other cases like remove-machine, continue to fail if the machine has
containers. In that case force can be used.