Rename suites/backup_restore to suites/backup to avoid underscore
juju-qa-jenkins/jobs/ci-run/scripts/lxd-runner.sh creates the container
name as follows:

CONTAINER_NAME="ci-build-${{SHORT_GIT_COMMIT}}-${{JOB_NAME}}"

And because the job name is created directly from the juju/tests/suites
subdirectory, the job name is "backup_restore".

lxc doesn't like underscores in container names, so the run fails.

Just getting rid of the underscore in the directory name seems like the
simplest way to fix.

This commit adds "-e" to the "echo" commands to fix the tests/main.sh
output for -h and -H (list of suites). Now that we're using bash instead
of sh we need "echo -e" to interpret the backslashes.