Remove transitive dependency to the apiserver package.
Some of the state tests imported api/testing just to access a bunch of
test helpers for creating and comparing macaroons. Unfortunately, the
api/testing package imports apiserver thus adding apiserver as a
transitive dependency to the state package. Consequently, running the
tests in the state package involved a needless full compilation step for
the apiserver package.

This commit copies the required versions into a dedicated file
(macaroon_helpers_test.go) into state and removes the transitive
dependency to apiserver.