Fix a race in the worker/introspection tests TestUnitStatusTimeout
What was happening is that sometimes (specifically in -race mode), the
Advance() call in the test was happening before the select on
time.After() in the worker code, and the time.After() never triggered.
Fixed by using WaitAdvance() -- see
https://github.com/juju/juju/wiki/Intermittent-failures

Additionally, there was no timeout on the testing http.Client, so the
Get() never returned and the test didn't stop till the 40 minute overall
timeout kicked in. Once I added an http.Client timeout the actual issue
above was a lot more clear.