Allow tools metadata generation to fail early
The metadata generation help documentation could be a lot more useful.
Previously you had to dig into the code to see the expected requirements
for the generation of the streams metadata from the tgz file.

Additionaly for tests you REALLY do not want to go look in the public
clouds for the streams if you're testing locally. The preventFallback
prevents the generation of tools to go off and do that.