Use metadata generation
The following removes the custom metadata testing server in favour of
the the generating metadata tool that we already have. The issue is that
the metadata tool also didn't do what I wanted, which was to fail hard
when I was wrong.

Failing hard now means that the tests will fail a lot harder a lot
sooner and tell you why at the right time, rather than when you attempt
to bootstrap! The reason is that the tool failed to do the right thing
and now you have to work that out yourself.