Upgrade controller
The following test ensures that we can upgrade our controller using
simplestreams from a known stable branch to the new juju.

There is a LOT of horse play here to get what we want to do. Namely that
we need to find a good stable version, luckily we can use
`--channel=<last version>/stable when installing a snap and that'll
give us a clean stable.

Then we need to perform more magic, as between major releases (2.8 and
2.9) we bumped the series version. So before we generate our metadata
agents we also need to create tar.gz files for both focal and bionic.
That way we can upgrade correctly, as 2.8 will expect that we can
upgrade to 2.9 bionic, even if they're the same binary!

Another magic trick is that we bootstrap with the new server url even
though we're not using that, but we fallback to the canonical
simplestreams for selecting the agent binary - magic /s

In addition to this we then do the normal things of upgrading the
controller and waiting for it to correctly upgrade.

Additional work for deploying should be added to ensure that we didn't
break things else where.