Starting to add lambda support
The following is a very initial idea for adding lambda support to for
each queries. It has not been generalised yet, but essentially it should
be doable.

The idea is that we have a series of scopes that a scope can add to when
we enter a lambda. This is kind of like stack registers, but not as
complicated. The reason we don't need to implement stack registers is
because we're using the go runtime to do it for us, so we can piggy back
on to that without having to do any of the work!

Additional work to change Ord to Box. The Box value is more generalised
and not only talks about Ord syntax, but syntax we need to do other
things like accessor values.

The idea of Box is the same as Java, except there is no bottom type.