Use functions for lambdas
Previously I was aiming to do `applications.forEach`, but actually that
introduces a lot of complexity in to the query evaluator, so much so it
makes it really hard to understand. Instead nothing can have a method
and we use high order functions to do this for us. We don't get partial
application, but I don't think we need it tbh.

The code is rather simple, lift the sub entity into it's own scope and
then use that when we call the lambda. We get a lot of things for free
and it works suprisingly well.

The accessor is identical to the infix in terms of using the order
precedence, but we have a new AST type so that the query evaluator
doesn't become a mess.

    juju-wait-for model default --query='forEach(applications, app =>
    app.status == "active")'

As there is no order in the forEach, it might be wise to order them to
get some valid expected behaviour.