Slices are non comparable
Attempting to use equality operator for all model config values are no
longer possible as we pass back a slice of interfaces. The way to do
this is then to use reflect.DeepEquals once we do know what the type is.
That then allows us not to take the hit for other values that are in the
model config.