Ensure mode doesn't have a default
You may ask why we shouldn't have a default for mode, well let me tell
you! When you use a different version that is prior to this commit, you
will end up throwing an error from here to eternity and by jove you can
get rid of it!

So the correct way to do this is to hide the mode, eventually when Juju
next (3 or 4) comes out we can set a default and that default will be
[]interface{}.

As we're introducing a new model config type schema.List(schema.String)
we need to ensure that we can handle this nicely.

The QA steps for this are rather simple...

    git checkout develp # 2.9 where the PR lands
    make go-install
    juju bootstrap lxd test
    juju model-config mode="[strict]"

Test in strict mode, then when you're done and you want to use an older
controller...

    juju model-config --reset mode # --reset - THIS IS IMPORTANT!

    git checkout 2.8
    juju destroy-controller test

Obviously it should be mentioned that you don't need any of this song
and dance when you're staying with the same client.