Goal state yaml
This introduces a yaml file as a goal state. It constructs a query from
a given yaml for the state of a deployment. The yaml is restricted on
purpose as it becomes harder to implement more logical operators and if
type statements. Instead a sub set of features are provided.

The code just reads the yaml file and constructs a query using the new
build nodes. It's rather easy then to construct any of the query DSL
from these nodes.