Force the revaluation of a model
The order of AllWatcher deltas might not be in an expected order, to
which the derived status of the application might not have been
processed if the unit delta hasn't arrived.

To fix this cache the model and force the revaluation every time, which
is worse for performance, but because it's on the client side, it
shouldn't matter too much. Yet still this is another time that the
application status change has forced client side changes.