Unfortunately yaml and json differ in serialization.
I'm pretty sure this was not intended, but we are using these structures
to format JSON and YAML content in 'network-get' output. And because the
structures didn't have YAML tags, they were getting bad serialization.
For now, I'm enshrining the bad serialization, because we don't want to
break compatibility arbitrarily. But I did want to highlight it instead
of having it be accidental, and then have us break it if we renamed a
field.