Consistency between output
The following updates the messages to the user about which channel a
charm is being read from. Previously we where displaying it in various
different ways and that's just not good practice.

Instead we use the suffix `from channel %s` and drop in the channel at
the end. I think this is good practice and farily well standardised now
that it should become useful to a user.