Bring in the updated github.com/juju/txn
It had a bug where if it got an error while removing transactions that
would bubble up into a panic because it wasn't properly handling the
error state.

This also brings in updated "lru" as it was a transitive dependency, but
I don't think that is a problem for our purposes.