Initial re-working of api opener strategy
Instead of the old legacy code for opening up a connection to the API we
drop the retry strategy loop as this is old code from when controllers
committed suicide and left machines around for them to try and
re-connect again. That code has been removed and so should this, we
should fail fast and early without retrying.

The strategy code is really simple, either connect with the current
password or fallback to the old password.

Note: this is a WIP and as so tests will need to be re-added to sure the
new code working.