Prevent fetching of oci timing out snap install
The following wraps the installing of the oci image in a timeout. As we
don't actually require the image to be installed when we bootstrap we
can be a bit fast and loose about it. If you don't have timeout
installed (BSD and macosx) then we exit and install the image using
manual intervention/bootstrap.

The code ensures that if it takes longer than we expect we just cancel
it, but importantly ignore the timeout error and return true. That way
we can still install the snap.

I've used 2 minutes because it can take a long time to install a snap on
slower systems, so we don't want to gobble that time up installing an
image that might not install anyway.