Fix intermittent fail in TestWatchPorts with clock.Advance timing issue
This is happening sometimes (quite rarely!) because the test clock's
Advance method (called by the test) may sometimes be called before the
hub watcher's TxnWatcher.loop() method has called w.clock.After(d)
(state/watcher/txnwatcher.go:231).

Fix is to add a test hook that we can wait for (some of the machinery
is already there, with the existing txnWatcherStarting event published).
Ideally we wouldn't do this by patching a global, but it's significantly
simpler as there are a lot of layers to wire something through to the
TxnWatcher, and this patch pattern is already used in hubwatcher.go so
there's definitely precedent.

Note: this test is called TestWatchMachinePorts in the develop branch.

For an example, see https://jenkins.juju.canonical.com/job/Unit-RunUnitTests-centos7/76/
which shows the following:

[LOG] 0:00.337 DEBUG juju.clock advancing a clock that has nothing waiting: cf. https://github.com/juju/juju/wiki/Intermittent-failures
machine_ports_test.go:436:
    wc.AssertChange(expectChange)
testing/watcher.go:234:
    c.Fatalf("watcher did not send change")
... Error: watcher did not send change