Subshell handling
This one is rather complicated, but essentially we can't run all the
commands inside the timeout, as the timeout command is in a subshell
that might not have access to all the parents env vars. This can cause
us not to find microk8s as it doesn't know about unique $PATH setup.

The fix for this is to then collect all the commands early, which we
know won't be expensive to run and then push the timeout lower to just
wrap the downloading of the images and if that fails we don't really
care.

This is the way.