Handle user charmstore urls
The following ensures that we correctly handle when there is no prefix
for a charmstore url. The way the url path is handled in the library
means we need to do a bit more work up front if we want to maintain
strict compatibility. This was not the original goal of the library as
the feature flag was intended to be lifted by now, but that's how it is.