Always request MAAS NIC in the default space even if it is not bound
The changes in the previous commit included a check that would ignore
the space for the default application binding. The reasoning was that if
all endpoints were explicitly bound to other spaces, there was no need
to provision a NIC in the default space since it was not used.

Based on the review feedback, this check seems too conservative and can
be relaxed so that we always request a NIC in the default space. This
will also allow users to rebind one or more of the application endpoints
to the default space.