Unicode support in info command
One of the missing features was adding unicode support to the info
command so that when you supplied info it would naturally add the
correct types in the output.

The following code adds that for info. If required we can move it to the
output folder if others find it more useful. It's basic in its primative
form, but does exactly what we want by identifying a location in the
code to print a unicode character out.

By default it is set to "auto" which is the same as snap cli and that it
will work out if your system can support it. It seems rather sane so
I've lifted in from there as they don't expose that function for
detection.

Unfortunately we don't have the same CLI structure to lift more of the
code which is a shame, but this is the way.