Charmhub not found errors
The following exposes the not found errors for a given request. There
are other errors we could be exposing (403, 500 for example) and showing
to the user.

This is the first pass that we do that and in a subsequent PR we can
handle that by exposing an error message to the user if required
explaining that the charm/bundle could not be found.

The code is simple, we return a RESTResponse that has a status code int
that we can check against the http.StatusCode before returning a
different error based on the the Error from the api response. It's a bit
convoluted because the status code is in the header and the message is
in a json blob.