Update status to show charmhub source
The following was noticed when testing an integration test and was noted
that charmhub charms where left as "unknown" and charmstore charms where
rendered as "jujucharms".

Jujucharms was an old domain name that now redirects to jaas.ai. Rather
than use domain names for the source, we should identify the store they
come from[1].

 - cs: is charmstore
 - ch: is charmhub
 - local: is local

Anything else is then uses "unknown".

The code is simple, but the tests are not as we have to sort out all the
status internal tests which are a pain.

There could be an underlying issue in processCommonModelApplicationArgs,
this expects that we either pass in a series or it gets it from the
charm url. This will never work for charmhub as the series is never
rendered for a charmhub url. I suspect a follow up PR to handle this
case is required.

 1. jujucharms.com -> jaas.ai -> charmhub.io (I'm sure I'm missing
 another as well), we should just not play this game.