Move DefaultSupportedLTS per juju release
Juju should be locked stepped with the default supported LTS and not
depend on an external package. This forces Juju to understand that when
a new LTS is updated and that we depend on that LTS then it should be
correctly updated when we do a release.

Unfortunately we missed the default LTS for 2.8 and are instead doing it
in 2.9, with these changes we should be in a better chance to spot that.

This PR does not update the default from bionic to focal, that will come
in a subsequent PR.

Cherry-picked from commit on develop:
https://github.com/juju/juju/commit/f2d957a56c73f943fe88590fdf076c35545d10fa