Rework dependency management in cmd
The bundle diff has a very bad anti-pattern in it, where we end up
leaving test code inside the command in production code. The if
statement might not be dangerous in terms of execution, but it can be
dangerous in terms of understanding the code and how you read it.

Instead we pass in different functions (factories) that can give the
command what it wants at the right time and not related to the if
command running at the right time.

The tests update and reflect the changes.

Also I would love to use mocks here!