Test filtering logic
The following updates the filtering logic for both arch and series
ensuring that we get the right channel when displaying info.

The logic is actually simple if we do some normalisation before we go
any further. That step includes swapping out empty space to mean "all".
With this default we can say that if any architecture and series match
either the specified series or the all keyword add it to the witnessed
set and return that back.

The other changes to info was introducing architecture and the fallout
from that.